WorkflowManagerをCLIControllerに名称変更/CLIControllerに#getIdメソッドを追加、readNote等のメソッドをprivateメソッドに変更/main.jsのexec()ファンクションを削除し、CLIControllerクラスにexecメソッドとして定義
